use uint64

+++ internal/common/hash.go
@@ -27,13 +27,13 @@ func WriteAndReturnSHA256(destination io.Writer, source io.Reader) (string, erro
 // it is a good fit for n-quads where we care about the data
 // but not the order of the quads inside
 type SumWriter struct {
-	Sum uint32
+	Sum uint64
 }
 
 // Write implements the io.Writer interface
 func (sw *SumWriter) Write(p []byte) (n int, err error) {
 	for _, b := range p {
-		sw.Sum += uint32(b)
+		sw.Sum += uint64(b)
 	}
 	return len(p), nil
 }
@@ -42,10 +42,10 @@ func (sw *SumWriter) ToString() string {
 	return fmt.Sprintf("%d", sw.Sum)
 }
 
-func ByteSum(b []byte) uint32 {
-	var sum uint32
+func ByteSum(b []byte) uint64 {
+	var sum uint64
 	for _, v := range b {
-		sum += uint32(v)
+		sum += uint64(v)
 	}
 	return sum
 }
